tests: remove fincore SMALLER_THAN_PAGESIZE tests
These tests failed for example on ext4 with a non-default
blocksize 1024. In this case it seems that the kernel only
caches a full page if we read at least PAGE_SIZE -BLOCK_SIZE +1.

Both tests also failed on NFS allthough block size seems to be
1M there.

We still keep the test for PAGE_SIZE-1 which seems to work more
reliable.

Signed-off-by: Ruediger Meier <ruediger.meier@ga-group.nl>