script: resolve the ambiguity about the optional argument of the -t option
The 'timing' option accepts an optional argument, but there is
a difference on the format of the argument accepted by the short and the
long version of the option: the short form expects the argument directly
after the '-t', while the long form '--timing' expects the file name
after an equal sign.

Document that to avoid confusing the user and have they pass the equal
sign also when using the short form which would result in a file name
starting with the equal character.

See https://github.com/karelzak/util-linux/issues/463

Signed-off-by: Antonio Ospite <ao2@ao2.it>