script: write typescript to filedescriptor
This change introduces the possibility to write the typscript data to a
file descriptor instead to a file.
This is useful if you want to implement a program which directly consumes
the data of the script util.

The new feature can be tested by executing the following procedure

make script
./script -h
echo "GOTO SCREEN,PRESS 'CTRL + d' to exit"
./script -d -e -f -q -t 5 \
      5> >(cat > test.typescript)  \
      2> >(cat > test.timing)
cat test.typescript |sed '~s,^,TYPESCRIPT |,'
cat test.timing|sed '~s,^,TIMING |,'